// src/model/history/earth/data-source/types.ts
export interface RawNationReference {
	tag: string
	name: string
	color: [number, number, number]
	graphicalCulture: string
	initialGovernmentType: string | null
	primaryCulture: string | null
	religion: string | null
}